Added whenever gem and scheduled task

+++ lib/tasks/trades.rake
@@ -3,8 +3,12 @@ namespace :trades do
   task clean: :environment do
     puts "Starting task"
     incomplete_trades = Trade.where(receiver_name: nil)
-    incomplete_trades.each { |t| puts "Removing trade for #{t.item }" }
-    incomplete_trades.destroy_all
+    if incomplete_trades.empty?
+      puts "No incomplete invitations"
+    else
+      incomplete_trades.each { |t| puts "Removing trade for #{t.item }" }
+      incomplete_trades.destroy_all
+    end
     puts "Done!"
   end
 end
